update api calls
