+++ components/form-builder/hooks/index.ts
@@ -1,10 +1,10 @@
 export { useAllowPublish } from "./useAllowPublish";
 export { useDeleteForm } from "./useDelete";
 export { useElementOptions } from "./useElementOptions";
-export { usePublish } from "./usePublish";
+export { useTemplateApi } from "./useTemplateApi";
 export { useActivePathname } from "./useActivePathname";
 export { useTemplateStatus } from "./useTemplateStatus";
-export { TemplateApiProvider, useTemplateApi } from "./useTemplateApi";
+export { TemplateApiProvider, useTemplateContext } from "./useTemplateContext";
 export { useIsWithin } from "./useIsWithin";
 export { usePanelActions } from "./usePanelActions";
 export { useUpdateElement } from "./useUpdateElement";
