+++ components/form-builder/hooks/useTemplateApi.tsx
@@ -1,55 +1,84 @@
-import React, { createContext, useState, useContext } from "react";
-import { useTemplateStore } from "../store";
-import { usePublish } from "../hooks";
-import { useTranslation } from "next-i18next";
-import { logMessage } from "@lib/logger";
+import { DeliveryOption, SecurityAttribute } from "@lib/types";
+import axios, { AxiosError } from "axios";
 
-interface TemplateApiType {
-  error: string | null;
-  saveForm: () => Promise<string | false>;
-}
-
-const defaultTemplateApi: TemplateApiType = {
-  error: null,
-  saveForm: async () => new Promise((resolve) => resolve(false)),
+type TemplateConfig = {
+  deliveryOption?: DeliveryOption;
+  securityAttribute?: SecurityAttribute;
 };
 
-const TemplateApiContext = createContext<TemplateApiType>(defaultTemplateApi);
-
-export function TemplateApiProvider({ children }: { children: React.ReactNode }) {
-  const { t } = useTranslation(["form-builder"]);
-  const [error, setError] = useState<string | null>(null);
-  const { id, getSchema, getName, getDeliveryOption } = useTemplateStore((s) => ({
-    id: s.id,
-    getSchema: s.getSchema,
-    getName: s.getName,
-    getDeliveryOption: s.getDeliveryOption,
-  }));
-
-  const { uploadJson } = usePublish();
-
-  const saveForm = async () => {
+export const useTemplateApi = () => {
+  const save = async ({
+    jsonConfig,
+    name,
+    formID,
+    publish = false,
+    deliveryOption,
+    securityAttribute,
+  }: {
+    jsonConfig: string;
+    name?: string;
+    formID?: string;
+    publish?: boolean;
+    deliveryOption?: DeliveryOption;
+    securityAttribute?: SecurityAttribute;
+  }) => {
+    let formData;
     try {
-      const result = await uploadJson(getSchema(), getName(), getDeliveryOption(), id);
+      formData = JSON.parse(jsonConfig);
+    } catch (e) {
+      if (e instanceof SyntaxError) {
+        return { error: new Error("failed to parse form data") };
+      }
+    }
 
-      if (result && result?.error) {
-        throw result?.error as Error;
+    try {
+      const url = formID ? `/api/templates/${formID}` : "/api/templates";
+
+      let data: TemplateConfig = {};
+
+      if (deliveryOption) {
+        data = { deliveryOption };
       }
 
-      setError(null);
-      return result?.id;
+      if (securityAttribute) {
+        data = { securityAttribute };
+      }
+
+      const result = await axios({
+        url: url,
+        method: formID ? "PUT" : "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        data: {
+          isPublished: publish ? true : false,
+          formConfig: formData,
+          name: name,
+          ...data,
+        },
+        timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
+      });
+
+      // PUT request with a { formID, isPublished } payload will update the Form Template isPublished property
+      if (publish && formID) {
+        await axios({
+          url: url,
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          data: {
+            isPublished: true,
+          },
+          timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
+        });
+      }
+
+      return { id: result?.data?.id };
     } catch (err) {
-      logMessage.error(err as Error);
-      setError(t("errorSaving"));
-      return false;
+      return { error: err as AxiosError };
     }
   };
 
-  return (
-    <TemplateApiContext.Provider value={{ error, saveForm }}>
-      {children}
-    </TemplateApiContext.Provider>
-  );
-}
-
-export const useTemplateApi = () => useContext(TemplateApiContext);
+  return { save };
+};
